+++ UI/Account Windows and Associated/EnterCashAmountWindow.xaml.cs
@@ -9,7 +9,7 @@ namespace Account_Windows
 	/// </summary>
 	public partial class EnterCashAmountWindow : Window
 	{
-		public double amount = 0;
+		public decimal amount = 0;
 
 		/// <summary>
 		/// Проверяет, является ли введенная строка корректным числом с плав. запятой
@@ -17,7 +17,7 @@ namespace Account_Windows
 		/// <param name="input">Введенная строка</param>
 		/// <param name="tmp">Преобразованное значение. Если ввод некорректный, то значение неопределено</param>
 		/// <returns>true/false. Если true, то в tmp результат преобразования</returns>
-		private bool IsInputValid(string input, out double tmp)
+		private bool IsInputValid(string input, out decimal tmp)
 		{
 			if (String.IsNullOrEmpty(input))
 			{
@@ -32,7 +32,7 @@ namespace Account_Windows
 				return false;
 			}
 
-			if (!Double.TryParse(input, out tmp))
+			if (!decimal.TryParse(input, out tmp))
 			{
 				MessageBox.Show("Некорректрый ввод! Введите число.");
 				Dispatcher.BeginInvoke((ThreadStart)delegate
@@ -73,7 +73,7 @@ namespace Account_Windows
 
 		private void btnOk_Click(object sender, RoutedEventArgs e)
 		{
-			if (IsInputValid(AmountEnterBox.Text, out double tmp))
+			if (IsInputValid(AmountEnterBox.Text, out decimal tmp))
 			{
 				amount = tmp;
 				DialogResult = true;
